Move RemoteClusterService into TransportService
TransportService and RemoteClusterService are closely coupled already today
and to simplify remote cluster integration down the road it can be a direct
dependency of TransportService. This change moves RemoteClusterService into
TransportService with the goal to make it a hidden implmementation detail
of TransportSerivce in followup changes.
