+++ test/framework/src/main/java/org/elasticsearch/test/transport/MockTransportService.java
@@ -819,7 +819,7 @@ public final class MockTransportService extends TransportService {
     }
 
     @Override
-    protected void doClose() {
+    protected void doClose() throws IOException {
         super.doClose();
         synchronized (openConnections) {
             assert openConnections.size() == 0 : "still open connections: " + openConnections;
